requirements.txt
python-telegram-bot==20.7
qrcode[pil]==7.4.2
Pillow==10.1.0
flask==3.0.0
gunicorn==21.2.0
